create the proflie page realtes #53 and make its private and make the add posts query close #53
