+++ private/style/profile.css
@@ -1,0 +1,113 @@
+.profile-container {
+  width: 70%;
+  height: 90vh;
+  padding-top: 50px;
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 2rem;
+}
+.profile-container .create-post {
+  height: 100%;
+  flex: 2;
+}
+.profile-container .create-post h2 {
+  font-size: 18px;
+  width: 100%;
+  padding-bottom: 20px;
+  border-bottom: 1px solid #fff;
+  font-weight: 400;
+}
+.profile-container .create-post .create-post-body {
+  margin-top: 20px;
+  width: 100%;
+  height: 80%;
+  background-color: #fff;
+  border-radius: 4px;
+  border: 1px solid var(--line);
+}
+.profile-container .create-post .create-post-body form {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+  flex-direction: column;
+  padding: 2rem;
+  gap: 1rem;
+}
+.profile-container .create-post .create-post-body form input {
+  height: 55px;
+  width: 100%;
+  padding: 0px 1rem;
+  border: 1px solid var(--line);
+  color: var(--placeHolder);
+  background-color: var(--field);
+  outline: none;
+  border-radius: 4px;
+}
+.profile-container .create-post .create-post-body form textarea {
+  height: 100%;
+  width: 100%;
+  padding: 1rem;
+  border: 1px solid var(--line);
+  color: var(--placeHolder);
+  background-color: var(--field);
+  outline: none;
+  border-radius: 4px;
+  resize: none;
+}
+.profile-container .create-post .create-post-body form button {
+  width: 100px;
+  padding: 10px 0;
+  align-self: flex-end;
+  outline: none;
+  border: none;
+  background-color: var(--button);
+  color: #fff;
+  font-size: 18px;
+  font-weight: 500;
+  border-radius: 4px;
+  cursor: pointer;
+}
+.profile-container .create-post .create-post-body form input:focus,
+.profile-container .create-post .create-post-body form input:hover {
+  border: 1px solid var(--button);
+}
+.profile-container .create-post .create-post-body form textarea:focus,
+.profile-container .create-post .create-post-body form input:focus,
+.profile-container .create-post .create-post-body form textarea:hover,
+.profile-container .create-post .create-post-body form input:hover {
+  border: 1px solid var(--button);
+}
+.profile-container .rules {
+  height: 300px;
+  align-self: flex-start;
+  flex: 0.9;
+  background-color: #fff;
+  border-radius: 4px;
+}
+.profile-container .rules h2 {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  gap: 4px;
+  padding: 10px;
+  font-size: 17px;
+  font-weight: 500;
+  width: 90%;
+  margin: 0 auto;
+  border-bottom: 1px solid var(--line);
+}
+.profile-container .rules ol {
+  list-style-position: inside;
+  padding: 15px;
+  font-size: 15px;
+  font-weight: 500;
+}
+.profile-container .rules ol li {
+  padding: 10px 0;
+  padding-left: 5px;
+  border-bottom: 1px solid var(--line);
+}
