+++ public/index.html
@@ -118,6 +118,14 @@
             <div class="main-body">
                 <section class="popular">
                     <div class="popular-body">
+                        <div class="p-header add-post" style="margin-bottom: 1rem;">
+                            <div class="add-post-img"><img
+                                    src="https://styles.redditmedia.com/t5_6zhglp/styles/profileIcon_snoo373b8fef-17fa-474c-882d-15f53e5ae3bd-headshot.png?width=256&height=256&crop=256:256,smart&s=8abaf96ecfbe41bdedf9b5a641c4bc79824cfa9c"
+                                    alt=""></div>
+                            <input type="text" placeholder="Create post">
+                            <div class="add-post-icons"><i class="fa-regular fa-image"></i> <i
+                                    class="fa-solid fa-link"></i></div>
+                        </div>
                         <div class="p-header">
                             <a href=""><i class="fa-solid fa-fire"></i> Hot</a>
                             <a href="">Everywhere <i class="fa-solid fa-chevron-down"></i></a>
@@ -127,8 +135,6 @@
                                     class="fa-regular fa-square-minus"></i> <i class="fa-solid fa-angle-down"></i></a>
                         </div>
                         <div class="p-posts">
-                            <!-- <i class="fa-solid fa-circle-up"></i> -->
-                            <!-- <i class="fa-solid fa-circle-down"></i> -->
                         </div>
                     </div>
                 </section>
