+++ server/controller/posts/addNewPost.js
@@ -1,6 +1,11 @@
+const { addNewPostQ } = require("../../database/queries");
+
 const addNewPost = (req, res) => {
-    //addNewPost query
-  };
-  
-  module.exports = addNewPost;
-  
+  const user_id = req.user.id;
+  const { title, content } = req.body;
+  addNewPostQ({ title, content, user_id })
+    .then((data) => res.send({ msg: "post added" }))
+    .catch((err) => res.send({ error: "post not added" }));
+};
+
+module.exports = addNewPost;
