+++ server/database/queries/posts/index.js
@@ -1,3 +1,4 @@
 const getAllPostsQ = require("./getAllPostsQ");
+const addNewPostQ = require("./addNewPostQ");
 
-module.exports = { getAllPostsQ };
+module.exports = { getAllPostsQ, addNewPostQ };
