+++ server/middlewares/index.js
@@ -1,3 +1,4 @@
 const isLogged = require("./isLogged");
+const virefyLogin = require("./virefyLogin");
 
-module.exports = { isLogged };
+module.exports = { isLogged, virefyLogin };
